Begin work on performance gains (and pretty code).

+++ cave_core/models/sessions.py
@@ -7,11 +7,17 @@ from django.utils.translation import gettext_lazy as _
 # Internal Imports
 from cave_api.api import execute_command
 from cave_core.models.cache import cache
-from cave_core.utils.constants import api_keys, background_api_keys
+from cave_core.utils.constants import (
+    background_api_keys,
+    api_keys_set,
+    background_api_keys_set,
+)
 from cave_core.websockets.cave_ws_broadcaster import CaveWSBroadcaster
 from cave_utils import Validator
 
 # External Imports
+import hashlib
+import orjson
 from pamda import pamda
 
 
@@ -91,6 +97,9 @@ class Sessions(models.Model):
             else:
                 cache.set(f"session:{self.id}:executing", False)
                 self.__dict__["is_executing"] = False
+                self.__dict__.pop("data", None)
+                self.__dict__.pop("versions", None)
+                self.__dict__.pop("hashes", None)
                 self.broadcast_loading(False)
 
     def get_user_ids(self) -> list:
@@ -137,7 +146,7 @@ class Sessions(models.Model):
         # Used a local object cached versions object to prevent multiple calls to the cache
         versions = self.__dict__.get("versions")
         if self.__dict__.get("is_executing") and versions:
-            return versions
+            return dict(versions)
         self.__dict__["versions"] = cache.get(f"session:{self.id}:versions", {})
         return dict(self.__dict__["versions"])
 
@@ -192,14 +201,13 @@ class Sessions(models.Model):
         if keys == None:
             keys = list(self.get_versions().keys())
         if client_only:
-            keys = pamda.intersection(keys, api_keys)
+            keys = [k for k in keys if k in api_keys_set]
         if len(omit_keys) > 0:
-            keys = pamda.difference(keys, omit_keys)
-        keys_to_get_from_cache = []
-        for key in keys:
-            # Avoid additional cache hits by checking if the data is already in the session __dict__
-            if not pamda.hasPath(path=["data", key], data=self.__dict__):
-                keys_to_get_from_cache.append(key)
+            omit_keys_set = set(omit_keys)
+            keys = [k for k in keys if k not in omit_keys_set]
+        data_cache = self.__dict__.setdefault("data", {})
+        hashes = self.__dict__.setdefault("hashes", {})
+        keys_to_get_from_cache = [key for key in keys if key not in data_cache]
         # If there any keys to get from the cache, get them all at once and update the session __dict__
         if len(keys_to_get_from_cache) > 0:
             new_data = cache.get_many(
@@ -227,11 +235,13 @@ class Sessions(models.Model):
                 self.execute_api_command(command="init", broadcast_changes=True, command_keys=[])
                 # Raise an exception to stop the current execution whatever it may be.
                 raise Exception("The data error should now be fixed. Please try your action again.")
-            for key, value in new_data.items():
-                if value != None:
-                    # Update the local session __dict__ with the new data to prevent multiple cache hits later
-                    pamda.assocPath(path=["data", key], value=value, data=self.__dict__)
-        return {key: pamda.path(["data", key], self.__dict__) for key in keys}
+            data_cache.update(new_data)
+            for k, val in new_data.items():
+                if val is not None and k not in hashes:
+                    hashes[k] = hashlib.md5(
+                        orjson.dumps(val, default=str, option=orjson.OPT_NON_STR_KEYS)
+                    ).digest()
+        return {key: data_cache.get(key) for key in keys}
 
     def broadcast_changed_data(
         self, previous_versions: dict, broadcast_loading: bool = True, force_overwrite: bool = False
@@ -315,24 +325,43 @@ class Sessions(models.Model):
         """
         # print('==REPLACE DATA==')
         versions = self.get_versions()
+        data_cache = self.__dict__.setdefault("data", {})
+        hashes = self.__dict__.setdefault("hashes", {})
+        keys_to_delete = []
+
         if wipeExisting:
-            data_keys = list(data.keys())
-            keys_to_delete = pamda.difference(list(versions.keys()), data_keys)
-            cache.delete_many(
-                [f"session:{self.id}:data:{key}" for key in keys_to_delete],
-                memory=True,
-                persistent=True,
-            )
-            for key in keys_to_delete:
-                versions.pop(key, None)
-        # Update the cache with the new data
-        cache.set_many({f"session:{self.id}:data:{key}": value for key, value in data.items()})
-        # Store the new data locally in the session __dict__ to prevent multiple cache hits
+            data_keys = set(data.keys())
+            keys_to_delete = [k for k in versions if k not in data_keys]
+            if keys_to_delete:
+                cache.delete_many(
+                    [f"session:{self.id}:data:{key}" for key in keys_to_delete],
+                    memory=True,
+                    persistent=True,
+                )
+                for key in keys_to_delete:
+                    versions.pop(key, None)
+                    data_cache.pop(key, None)
+                    hashes.pop(key, None)
+
+        changed_data = {}
         for key, value in data.items():
-            pamda.assocPath(path=["data", key], value=value, data=self.__dict__)
-            versions[key] = versions.get(key, 0) + 1
-        # Update versions post replacement
-        self.set_versions(versions)
+            val_hash = (
+                hashlib.md5(
+                    orjson.dumps(value, default=str, option=orjson.OPT_NON_STR_KEYS)
+                ).digest()
+                if value is not None
+                else None
+            )
+            if key not in hashes or hashes[key] != val_hash:
+                changed_data[f"session:{self.id}:data:{key}"] = value
+                data_cache[key] = value
+                hashes[key] = val_hash
+                versions[key] = versions.get(key, 0) + 1
+
+        if changed_data or keys_to_delete:
+            if changed_data:
+                cache.set_many(changed_data)
+            self.set_versions(versions)
         # print('==REPLACE DATA END==')
 
     def execute_api_command(
@@ -381,9 +410,7 @@ class Sessions(models.Model):
             session_data=session_data, command=command, socket=socket, mutate_dict=mutate_dict
         )
         # Ensure that no reserved api keys are returned
-        background_api_keys_used = pamda.intersection(
-            list(command_output.keys()), background_api_keys
-        )
+        background_api_keys_used = [k for k in command_output if k in background_api_keys_set]
         if len(background_api_keys_used) > 0:
             raise Exception(
                 f"Oops! The following reserved api keys were returned: {str(background_api_keys_used)}"
@@ -515,11 +542,11 @@ class Sessions(models.Model):
             - What: The new session object that was created
         """
         session_data = self.get_data(keys=list(self.get_versions().keys()), client_only=False)
-        new_session = self
-        new_session.name = str(name)
-        new_session.description = str(description)
-        new_session.pk = None
-        new_session.save()
+        new_session = Sessions.objects.create(
+            name=str(name),
+            description=str(description),
+            team=self.team,
+        )
         cache.set_many(
             {f"session:{new_session.id}:data:{key}": value for key, value in session_data.items()}
         )
